Load algorithm improvements
The code now supports two algorithms, you can decide at compile time.
The first algorithm (the default) will turn on the load when the
battery is not fully discharged and there is no solar power. You'd use
that for lighting. The second algorithm turns on the load when the
battery is fully charged and there is solar power. You'd use that as
a dump circuit (if you have something to do with excess power).